@@ -77,56 +103,64 @@
       "packageCategory": "games",
       "depends": [
         "gawk",
-        "curl"
+        "curl",
+        "zstd"
       ]
     },
     "pacman": {
       "packageCategory": "games",
       "depends": [
         "gawk",
-        "curl"
+        "curl",
+        "zstd"
       ]
     },
     "rpm": {
       "packageCategory": "games",
       "depends": [
         "gawk",
-        "curl"
+        "curl",
+        "zstd"
       ]
     }
   },
   "dependencies": {
-    "@fontsource/cabin": "^4.2.1",
-    "@fontsource/rubik": "^4.2.1",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@fontsource/cabin": "^4.5.1",
+    "@fontsource/rubik": "^4.5.1",
     "@fortawesome/fontawesome-svg-core": "^1.3.0-beta3",
     "@fortawesome/free-brands-svg-icons": "^6.0.0-beta3",
     "@fortawesome/free-solid-svg-icons": "^6.0.0-beta3",
     "@fortawesome/react-fontawesome": "^0.1.16",
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
+    "@mui/icons-material": "^5.3.1",
+    "@mui/material": "^5.4.0",
     "axios": "^0.21.1",
     "classnames": "^2.2.6",
     "discord-rich-presence-typescript": "^0.0.8",
     "electron-store": "^8.0.1",
     "electron-updater": "^4.6.1",
     "graceful-fs": "^4.2.8",
-    "i18next": "^20.2.2",
-    "i18next-browser-languagedetector": "^6.1.1",
-    "i18next-fs-backend": "^1.1.1",
-    "i18next-http-backend": "^1.2.2",
-    "plist": "^3.0.4",
+    "heroic-wine-downloader": "^1.2.2",
+    "i18next": "^21.6.6",
+    "i18next-fs-backend": "^1.1.4",
+    "i18next-http-backend": "^1.3.1",
+    "ini": "^2.0.0",
+    "plist": "^3.0.5",
     "pretty-bytes": "^5.6.0",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-contextmenu": "^2.14.0",
-    "react-dom": "^17.0.2",
-    "react-i18next": "^11.8.15",
+    "react-dom": "^18.0.0",
+    "react-i18next": "^11.15.3",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "systeminformation": "^5.9.17",
+    "react-scripts": "^5.0.0",
+    "simple-keyboard": "^3.4.24",
+    "source-map-support": "^0.5.21",
+    "systeminformation": "^5.10.7",
     "tslib": "^2.3.1"
   },
   "scripts": {
-    "start": "nf start -p 3000 && electron-start",
+    "start": "DISABLE_ESLINT_PLUGIN=true nf start -p 3000 && electron-start",
     "electron-start": "ELECTRON_IS_DEV=1 node public/start-react",
     "electron": "yarn build-electron && electron . --trace-warnings",
     "react-start": "HOST=localhost BROWSER=none react-scripts start",
@@ -164,31 +200,28 @@
     "@testing-library/user-event": "^13.1.9",
     "@types/classnames": "^2.2.11",
     "@types/i18next-fs-backend": "^1.0.0",
+    "@types/ini": "^1.3.31",
     "@types/jest": "^26.0.23",
-    "@types/node": "^14.7.0",
+    "@types/node": "^17.0.10",
     "@types/plist": "^3.0.2",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
     "@types/react-router-dom": "^5.1.6",
-    "@typescript-eslint/eslint-plugin": "^4.26.0",
-    "@typescript-eslint/parser": "^4.26.0",
-    "electron": "15.3.4",
-    "electron-builder": "^22.14.5",
+    "@typescript-eslint/eslint-plugin": "^5.9.1",
+    "@typescript-eslint/parser": "^5.9.1",
+    "electron": "^18.0.1",
+    "electron-builder": "^22.14.13",
     "electron-devtools-installer": "^3.2.0",
-    "eslint": "^7.27.0",
+    "eslint": "^8.7.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-react": "^7.24.0",
-    "eslint-plugin-sort-exports": "^0.4.0",
-    "eslint-plugin-sort-imports-es6-autofix": "^0.6.0",
-    "eslint-plugin-sort-keys-fix": "^1.1.1",
-    "eslint-plugin-typescript-sort-keys": "^1.5.0",
     "foreman": "^3.0.1",
-    "husky": "^6.0.0",
-    "i18next-parser": "^4.2.0",
+    "husky": "^7.0.4",
+    "i18next-parser": "^5.4.0",
     "prettier": "^2.5.1",
     "pretty-quick": "^3.1.2",
-    "ts-jest": "^26.5.6",
-    "typescript": "^4.3.2"
+    "ts-jest": "^27.1.3",
+    "typescript": "^4.5.5"
   },
   "browserslist": {
     "production": [
