@@ -2,11 +2,11 @@ import './index.css'
 
 import React, { lazy, useContext, useEffect, useRef, useState } from 'react'
 
-import { GameInfo } from 'src/types'
+import { GameInfo, Runner } from 'src/types'
 import ContextProvider from 'src/state/ContextProvider'
 import cx from 'classnames'
 
-import ArrowDropUp from '@material-ui/icons/ArrowDropUp'
+import ArrowDropUp from '@mui/icons-material/ArrowDropUp'
 import { UpdateComponent } from 'src/components/UI'
 import { useTranslation } from 'react-i18next'
 import { getLibraryTitle } from './constants'
@@ -25,7 +25,11 @@ interface Props {
 export const Library = ({ library, showRecentsOnly }: Props) => {
   const { layout, gameUpdates, refreshing, category, filter } =
     useContext(ContextProvider)
-  const [showModal, setShowModal] = useState({ game: '', show: false })
+  const [showModal, setShowModal] = useState({
+    game: '',
+    show: false,
+    runner: 'legendary' as Runner
+  })
   const { t } = useTranslation()
   const backToTopElement = useRef(null)
 
@@ -52,12 +56,12 @@ export const Library = ({ library, showRecentsOnly }: Props) => {
     }
   }
 
-  function handleModal(appName: string) {
-    setShowModal({ game: appName, show: true })
+  function handleModal(appName: string, runner: Runner) {
+    setShowModal({ game: appName, show: true, runner })
   }
 
   if (refreshing && !showRecentsOnly) {
-    return <UpdateComponent />
+    return <UpdateComponent inline />
   }
 
   function titleWithIcons() {
@@ -74,7 +78,10 @@ export const Library = ({ library, showRecentsOnly }: Props) => {
       {showModal.show && (
         <InstallModal
           appName={showModal.game}
-          backdropClick={() => setShowModal({ game: '', show: false })}
+          runner={showModal.runner}
+          backdropClick={() =>
+            setShowModal({ game: '', show: false, runner: 'legendary' })
+          }
         />
       )}
       <h3 className="libraryHeader">
@@ -97,6 +104,8 @@ export const Library = ({ library, showRecentsOnly }: Props) => {
               app_name,
               is_installed,
               is_mac_native,
+              is_linux_native,
+              runner,
               is_game,
               install: { version, install_size, is_dlc }
             }: GameInfo) => {
@@ -107,6 +116,7 @@ export const Library = ({ library, showRecentsOnly }: Props) => {
               return (
                 <GameCard
                   key={app_name}
+                  runner={runner}
                   cover={art_square}
                   coverList={art_cover}
                   logo={art_logo}
@@ -117,9 +127,10 @@ export const Library = ({ library, showRecentsOnly }: Props) => {
                   version={`${version}`}
                   size={`${install_size}`}
                   hasUpdate={hasUpdate}
-                  buttonClick={() => handleModal(app_name)}
+                  buttonClick={() => handleModal(app_name, runner)}
                   forceCard={showRecentsOnly}
                   isMacNative={is_mac_native}
+                  isLinuxNative={is_linux_native}
                 />
               )
             }
