@@ -10,8 +10,8 @@ import ElectronStore from 'electron-store'
 import classNames from 'classnames'
 
 import { IpcRenderer } from 'electron'
-import Backspace from '@material-ui/icons/Backspace'
-import CreateNewFolder from '@material-ui/icons/CreateNewFolder'
+import Backspace from '@mui/icons-material/Backspace'
+import CreateNewFolder from '@mui/icons-material/CreateNewFolder'
 
 const { ipcRenderer } = window.require('electron') as {
   ipcRenderer: IpcRenderer
@@ -25,6 +25,7 @@ const storage: Storage = window.localStorage
 
 interface Props {
   altLegendaryBin: string
+  altGogdlBin: string
   checkForUpdatesOnStartup: boolean
   darkTrayIcon: boolean
   defaultInstallPath: string
@@ -34,17 +35,20 @@ interface Props {
   language: string
   maxWorkers: number
   showUnrealMarket: boolean
+  minimizeOnLaunch: boolean
   setDefaultInstallPath: (value: string) => void
   setEgsLinkedPath: (value: string) => void
   setEgsPath: (value: string) => void
   setLanguage: (value: string) => void
   setAltLegendaryBin: (value: string) => void
+  setAltGogdlBin: (value: string) => void
   setMaxWorkers: (value: number) => void
   startInTray: boolean
   toggleDarkTrayIcon: () => void
   toggleStartInTray: () => void
   toggleCheckUpdatesOnStartup: () => void
   toggleTray: () => void
+  toggleMinimizeOnLaunch: () => void
   toggleUnrealMarket: () => void
 }
 
@@ -56,6 +60,8 @@ export default function GeneralSettings({
   setEgsPath,
   altLegendaryBin,
   setAltLegendaryBin,
+  altGogdlBin,
+  setAltGogdlBin,
   egsLinkedPath,
   setEgsLinkedPath,
   showUnrealMarket,
@@ -70,17 +76,23 @@ export default function GeneralSettings({
   setMaxWorkers,
   darkTrayIcon,
   toggleDarkTrayIcon,
-  toggleCheckUpdatesOnStartup
+  toggleCheckUpdatesOnStartup,
+  toggleMinimizeOnLaunch,
+  minimizeOnLaunch
 }: Props) {
   const [isSyncing, setIsSyncing] = useState(false)
   const [maxCpus, setMaxCpus] = useState(maxWorkers)
   const [legendaryVersion, setLegendaryVersion] = useState('')
+  const [gogdlVersion, setGogdlVersion] = useState('')
   const { platform, refreshLibrary, isRTL } = useContext(ContextProvider)
   const { t, i18n } = useTranslation()
   const isLinked = Boolean(egsLinkedPath.length)
   const isWindows = platform === 'win32'
 
-  const settings = configStore.get('settings') as { altLeg: string }
+  const settings = configStore.get('settings') as {
+    altLeg: string
+    altGogdl: string
+  }
 
   useEffect(() => {
     i18n.changeLanguage(language)
@@ -90,7 +102,6 @@ export default function GeneralSettings({
   useEffect(() => {
     const getMoreInfo = async () => {
       const cores = await ipcRenderer.invoke('getMaxCpus')
-      const legendaryVer = await ipcRenderer.invoke('getLegendaryVersion')
       configStore.set('settings', {
         ...settings,
         altLeg: altLegendaryBin
@@ -98,6 +109,7 @@ export default function GeneralSettings({
 
       setMaxCpus(cores)
 
+      const legendaryVer = await ipcRenderer.invoke('getLegendaryVersion')
       if (legendaryVer === 'invalid') {
         setLegendaryVersion('Invalid')
         setTimeout(() => {
@@ -109,6 +121,26 @@ export default function GeneralSettings({
     }
     getMoreInfo()
   }, [maxWorkers, altLegendaryBin])
+
+  useEffect(() => {
+    const getGogdlVersion = async () => {
+      configStore.set('settings', {
+        ...settings,
+        altGogdl: altGogdlBin
+      })
+      const gogdlVersion = await ipcRenderer.invoke('getGogdlVersion')
+      if (gogdlVersion === 'invalid') {
+        setGogdlVersion('Invalid')
+        setTimeout(() => {
+          setAltGogdlBin('')
+          return setGogdlVersion('')
+        }, 3000)
+      }
+      return setGogdlVersion(gogdlVersion)
+    }
+
+    getGogdlVersion()
+  }, [altGogdlBin])
 
   async function handleSync() {
     setIsSyncing(true)
@@ -159,7 +191,7 @@ export default function GeneralSettings({
         properties: ['openDirectory'],
         title: t('box.choose-egs-prefix')
       })
-      .then(({ path }: Path) => setEgsPath(path ? `'${path}'` : ''))
+      .then(({ path }: Path) => setEgsPath(path ? path : ''))
   }
 
   function handleLegendaryBinary() {
@@ -172,7 +204,20 @@ export default function GeneralSettings({
           'Select Legendary Binary (needs restart)'
         )
       })
-      .then(({ path }: Path) => setAltLegendaryBin(path ? `'${path}'` : ''))
+      .then(({ path }: Path) => setAltLegendaryBin(path ? path : ''))
+  }
+
+  function handleGogdlBinary() {
+    return ipcRenderer
+      .invoke('openDialog', {
+        buttonLabel: t('box.choose'),
+        properties: ['openFile'],
+        title: t(
+          'box.choose-gogdl-binary',
+          'Select GOGDL Binary (needs restart)'
+        )
+      })
+      .then(({ path }: Path) => setAltGogdlBin(path ? path : ''))
   }
 
   async function handleChangeLanguage(language: string) {
@@ -224,7 +269,7 @@ export default function GeneralSettings({
                   title: t('box.default-install-path')
                 })
                 .then(({ path }: Path) =>
-                  setDefaultInstallPath(path ? `'${path}'` : defaultInstallPath)
+                  setDefaultInstallPath(path ? `${path}` : defaultInstallPath)
                 )
             }
             className="material-icons settings folder"
@@ -279,6 +324,54 @@ export default function GeneralSettings({
         <span className="smallMessage">
           {t('other.legendary-version', 'Legendary Version: ')}
           {legendaryVersion}
+        </span>
+      </span>
+      <span className="setting">
+        <span className={classNames('settingText', { isRTL: isRTL })}>
+          {t(
+            'setting.alt-gogdl-bin',
+            'Choose an Alternative GOGDL Binary to use (needs restart)'
+          )}
+        </span>
+        <span>
+          <input
+            data-testid="setting-alt-gogdl"
+            type="text"
+            placeholder={t(
+              'placeholder.alt-gogdl-bin',
+              'Using built-in GOGDL binary...'
+            )}
+            className="settingSelect"
+            value={altGogdlBin.replaceAll("'", '')}
+            onChange={(event) => setAltGogdlBin(event.target.value)}
+          />
+          {!altGogdlBin.length ? (
+            <SvgButton
+              onClick={() => handleGogdlBinary()}
+              className="material-icons settings folder"
+            >
+              <CreateNewFolder
+                data-testid="setGogdlBinaryButton"
+                style={{
+                  color: altGogdlBin.length ? 'transparent' : '#B0ABB6'
+                }}
+              />
+            </SvgButton>
+          ) : (
+            <SvgButton
+              className="material-icons settings folder"
+              onClick={() => setAltGogdlBin('')}
+            >
+              <Backspace
+                data-testid="setGogdlBinaryBackspace"
+                style={{ color: '#B0ABB6' }}
+              />
+            </SvgButton>
+          )}
+        </span>
+        <span className="smallMessage">
+          {t('other.gogdl-version', 'GOGDL Version: ')}
+          {gogdlVersion}
         </span>
       </span>
       {!isWindows && (
@@ -338,51 +431,70 @@ export default function GeneralSettings({
               }`}
             </button>
           </span>
+          {!isWindows && (
+            <InfoBox text="infobox.help">{t('help.general')}</InfoBox>
+          )}
         </span>
       )}
+
       {isWindows && (
         <span className="setting">
-          <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-            {t('setting.egs-sync')}
+          <label className={classNames('toggleWrapper', { isRTL: isRTL })}>
             <ToggleSwitch
               dataTestId="syncToggle"
               value={isLinked}
               handleChange={handleSync}
               title={t('setting.egs-sync')}
             />
-          </span>
+            <span>{t('setting.egs-sync')}</span>
+          </label>
         </span>
       )}
       <span className="setting">
-        <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-          {t('setting.exit-to-tray')}
+        <label className={classNames('toggleWrapper', { isRTL: isRTL })}>
           <ToggleSwitch
             dataTestId="exitToTray"
             value={exitToTray}
             handleChange={toggleTray}
             title={t('setting.exit-to-tray')}
           />
-        </span>
+          <span>{t('setting.exit-to-tray')}</span>
+        </label>
       </span>
       {exitToTray && (
         <span className="setting">
-          <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-            {t('setting.start-in-tray', 'Start Minimized')}
+          <label className={classNames('toggleWrapper', { isRTL: isRTL })}>
             <ToggleSwitch
               dataTestId="startInTray"
               value={startInTray}
               handleChange={toggleStartInTray}
               title={t('setting.start-in-tray', 'Start Minimized')}
             />
-          </span>
+            <span>{t('setting.start-in-tray', 'Start Minimized')}</span>
+          </label>
         </span>
       )}
       <span className="setting">
-        <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-          {t(
-            'setting.showUnrealMarket',
-            'Show Unreal Marketplace (needs restart)'
-          )}
+        <label className={classNames('toggleWrapper', { isRTL: isRTL })}>
+          <ToggleSwitch
+            dataTestId="minimizeOnLaunch"
+            value={minimizeOnLaunch}
+            handleChange={toggleMinimizeOnLaunch}
+            title={t(
+              'setting.minimize-on-launch',
+              'Minimize Heroic After Game Launch'
+            )}
+          />
+          <span>
+            {t(
+              'setting.minimize-on-launch',
+              'Minimize Heroic After Game Launch'
+            )}
+          </span>
+        </label>
+      </span>
+      <span className="setting">
+        <label className={classNames('toggleWrapper', { isRTL: isRTL })}>
           <ToggleSwitch
             value={showUnrealMarket}
             handleChange={() => toggleUnrealMarket()}
@@ -391,11 +503,16 @@ export default function GeneralSettings({
               'Show Unreal Marketplace (needs restart)'
             )}
           />
-        </span>
+          <span>
+            {t(
+              'setting.showUnrealMarket',
+              'Show Unreal Marketplace (needs restart)'
+            )}
+          </span>
+        </label>
       </span>
       <span className="setting">
-        <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-          {t('setting.darktray', 'Use Dark Tray Icon')}
+        <label className={classNames('toggleWrapper', { isRTL: isRTL })}>
           <ToggleSwitch
             value={darkTrayIcon}
             handleChange={() => {
@@ -404,14 +521,11 @@ export default function GeneralSettings({
             }}
             title={t('setting.darktray', 'Use Dark Tray Icon (needs restart)')}
           />
-        </span>
+          <span>{t('setting.darktray', 'Use Dark Tray Icon')}</span>
+        </label>
       </span>
       <span className="setting">
-        <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-          {t(
-            'setting.checkForUpdatesOnStartup',
-            'Check For Updates On Startup'
-          )}
+        <label className={classNames('toggleWrapper', { isRTL: isRTL })}>
           <ToggleSwitch
             value={checkForUpdatesOnStartup}
             handleChange={toggleCheckUpdatesOnStartup}
@@ -420,16 +534,21 @@ export default function GeneralSettings({
               'Check For Updates On Startup'
             )}
           />
-        </span>
+          <span>
+            {t(
+              'setting.checkForUpdatesOnStartup',
+              'Check For Updates On Startup'
+            )}
+          </span>
+        </label>
       </span>
       <span className="setting">
-        <span className={classNames('toggleWrapper', { isRTL: isRTL })}>
-          {t('setting.maxworkers')}
+        <label className={classNames('toggleWrapper', { isRTL: isRTL })}>
           <select
             data-testid="setMaxWorkers"
             onChange={(event) => setMaxWorkers(Number(event.target.value))}
             value={maxWorkers}
-            className="settingSelect smaller"
+            className="settingSelect smaller is-drop-down"
           >
             {Array.from(Array(maxCpus).keys()).map((n) => (
               <option key={n + 1}>{n + 1}</option>
@@ -438,9 +557,9 @@ export default function GeneralSettings({
               Max
             </option>
           </select>
-        </span>
+          <span>{t('setting.maxworkers')}</span>
+        </label>
       </span>
-      <InfoBox text="infobox.help">{t('help.general')}</InfoBox>
     </>
   )
 }
