@@ -5,17 +5,19 @@ import {
   GameSettings,
   InstallArgs,
   InstallInfo,
-  LaunchResult
+  LaunchResult,
+  Runner
 } from './types'
 
-type Runner = 'legendary' | 'gog'
 abstract class Game {
-  public static get(appName: string, runner: Runner = 'legendary') {
+  public static get(
+    appName: string,
+    runner: Runner = 'legendary'
+  ): LegendaryGame | GOGGame {
     if (runner === 'legendary') {
       return LegendaryGame.get(appName)
     } else if (runner === 'gog') {
-      logWarning('GOG integration is unimplemented.')
-      return null
+      return GOGGame.get(appName)
     }
   }
 
@@ -35,11 +37,11 @@ abstract class Game {
   abstract stop(): Promise<void>
   abstract syncSaves(arg: string, path: string): Promise<ExecResult>
   abstract uninstall(): Promise<ExecResult>
-  abstract update(): Promise<unknown>
+  abstract update(): Promise<{ status: 'done' | 'error' }>
 }
 
 import { LegendaryGame } from './legendary/games'
-import { logWarning } from './logger'
 import { BrowserWindow } from 'electron'
+import { GOGGame } from './gog/games'
 
 export { Game, Runner }
