@@ -1,6 +1,6 @@
 import { BrowserWindow, dialog } from 'electron'
-import { Game } from './games'
-import { logInfo } from './logger'
+import { Game, Runner } from './games'
+import { logInfo, LogPrefix } from './logger/logger'
 import i18next from 'i18next'
 
 export async function handleProtocol(window: BrowserWindow, url: string) {
@@ -9,23 +9,28 @@ export async function handleProtocol(window: BrowserWindow, url: string) {
   if (!url || scheme !== 'heroic' || !path) {
     return
   }
-  let [command, arg] = path?.split('/')
+  let [command, arg] = path.split('/')
   if (!command || !arg) {
     command = path
     arg = null
   }
-  logInfo(`ProtocolHandler: received '${url}'`)
+  logInfo(`received '${url}'`, LogPrefix.ProtocolHandler)
   if (command === 'ping') {
-    return logInfo('Received ping! Arg:', arg)
+    return logInfo(['Received ping! Arg:', arg], LogPrefix.ProtocolHandler)
   }
   if (command === 'launch') {
-    const game = Game.get(arg)
-    const { is_installed, title, app_name } = await game.getGameInfo()
+    let runner: Runner = 'legendary'
+    let game = await Game.get(arg, runner).getGameInfo()
+    if (!game) {
+      runner = 'gog'
+      game = await Game.get(arg, runner).getGameInfo()
+    }
+    const { is_installed, title, app_name } = game
     setTimeout(async () => {
       // wait for the frontend to be ready
       if (!is_installed) {
-        logInfo(`ProtocolHandler: "${arg}" not installed.`)
-        const { response } = await dialog.showMessageBox({
+        logInfo(`"${arg}" not installed.`, LogPrefix.ProtocolHandler)
+        const { response } = await dialog.showMessageBox(window, {
           buttons: [i18next.t('box.yes'), i18next.t('box.no')],
           cancelId: 1,
           message: `${title} ${i18next.t(
@@ -46,16 +51,17 @@ export async function handleProtocol(window: BrowserWindow, url: string) {
           if (filePaths[0]) {
             return window.webContents.send('installGame', {
               appName: app_name,
+              runner,
               installPath: filePaths[0]
             })
           }
         }
         if (response === 1) {
-          return logInfo('Not installing game')
+          return logInfo('Not installing game', LogPrefix.ProtocolHandler)
         }
       }
       mainWindow.hide()
-      window.webContents.send('launchGame', arg)
+      window.webContents.send('launchGame', arg, runner)
     }, 3000)
   }
 }
