@@ -1,21 +1,64 @@
 import * as axios from 'axios'
-import { app, dialog, net, shell } from 'electron'
+import { app, dialog, net, shell, Notification, BrowserWindow } from 'electron'
 import { exec } from 'child_process'
-import { existsSync, stat, rm } from 'graceful-fs'
+import { existsSync, rm, stat } from 'graceful-fs'
 import { promisify } from 'util'
-import i18next from 'i18next'
+import i18next, { t } from 'i18next'
 import prettyBytes from 'pretty-bytes'
 import si from 'systeminformation'
 import Store from 'electron-store'
 
 import { GlobalConfig } from './config'
-import { heroicGamesConfigPath, home, icon, legendaryBin } from './constants'
-import { logError, logInfo, logWarning } from './logger'
+import {
+  configStore,
+  fixAsarPath,
+  heroicConfigPath,
+  heroicGamesConfigPath,
+  icon,
+  isWindows
+} from './constants'
+import { logError, logInfo, LogPrefix, logWarning } from './logger/logger'
+import { basename, dirname, join } from 'path'
+import { runLegendaryCommand } from './legendary/library'
+import { runGogdlCommand } from './gog/library'
 
 const execAsync = promisify(exec)
 const statAsync = promisify(stat)
 
 const { showErrorBox, showMessageBox } = dialog
+
+export function showErrorBoxModal(
+  window: BrowserWindow | undefined | null,
+  title: string,
+  message: string
+) {
+  if (window) {
+    showMessageBox(window, {
+      type: 'error',
+      title,
+      message
+    })
+  } else {
+    showErrorBox(title, message)
+  }
+}
+
+export function showErrorBoxModalAuto(title: string, message: string) {
+  let window: BrowserWindow
+  try {
+    window = BrowserWindow.getFocusedWindow()
+    if (!window) {
+      window = BrowserWindow.getAllWindows()[0]
+    }
+  } catch (e) {
+    // empty
+  }
+  if (window) {
+    showErrorBoxModal(window, title, message)
+  } else {
+    showErrorBox(title, message)
+  }
+}
 
 /**
  * Compares 2 SemVer strings following "major.minor.patch".
@@ -37,27 +80,76 @@ async function isOnline() {
   return net.isOnline()
 }
 
-export const getLegendaryVersion = async () => {
-  const { altLegendaryBin } = await GlobalConfig.get().getSettings()
+async function isEpicServiceOffline(
+  type: 'Epic Games Store' | 'Fortnite' | 'Rocket League' = 'Epic Games Store'
+) {
+  const epicStatusApi = 'https://status.epicgames.com/api/v2/components.json'
+  const notification = new Notification({
+    title: `${type} ${t('epic.offline-notification-title', 'offline')}`,
+    body: t(
+      'epic.offline-notification-body',
+      'Heroic will maybe not work probably!'
+    ),
+    urgency: 'normal',
+    timeoutType: 'default',
+    silent: false
+  })
+
   try {
-    if (altLegendaryBin && !altLegendaryBin.includes('legendary')) {
-      return 'invalid'
+    const { data } = await axios.default.get(epicStatusApi)
+
+    for (const component of data.components) {
+      const { name: name, status: indicator } = component
+
+      // found component and checking status
+      if (name === type) {
+        const isOffline = indicator === 'major'
+        if (isOffline) {
+          notification.show()
+        }
+        return isOffline
+      }
     }
-    const { stdout } = await execAsync(`${legendaryBin} --version`)
-    return stdout
-      .split('legendary version')[1]
-      .replaceAll('"', '')
-      .replaceAll(', codename', '')
-      .replaceAll('\n', '')
+
+    notification.show()
+    return false
   } catch (error) {
+    logError(
+      `Failed to get epic service status with ${error}`,
+      LogPrefix.Backend
+    )
+    return false
+  }
+}
+
+export const getLegendaryVersion = async () => {
+  const { stdout, error } = await runLegendaryCommand(['--version'])
+
+  if (error) {
     return 'invalid'
   }
+
+  return stdout
+    .split('legendary version')[1]
+    .replaceAll('"', '')
+    .replaceAll(', codename', '')
+    .replaceAll('\n', '')
+}
+
+export const getGogdlVersion = async () => {
+  const { stdout, error } = await runGogdlCommand(['--version'])
+
+  if (error) {
+    return 'invalid'
+  }
+
+  return stdout
 }
 
 export const getHeroicVersion = () => {
   const VERSION_NUMBER = app.getVersion()
   const BETA_VERSION_NAME = 'Caesar Clown'
-  const STABLE_VERSION_NAME = 'Roronoa Zoro'
+  const STABLE_VERSION_NAME = 'Oden'
   const isBetaorAlpha =
     VERSION_NUMBER.includes('alpha') || VERSION_NUMBER.includes('beta')
   const VERSION_NAME = isBetaorAlpha ? BETA_VERSION_NAME : STABLE_VERSION_NAME
@@ -67,13 +159,13 @@ export const getHeroicVersion = () => {
 
 async function checkForUpdates() {
   const { checkForUpdatesOnStartup } = await GlobalConfig.get().getSettings()
-  logInfo('checking for heroic updates')
+  logInfo('checking for heroic updates', LogPrefix.Backend)
   if (!checkForUpdatesOnStartup) {
-    logInfo('skipping heroic updates')
+    logInfo('skipping heroic updates', LogPrefix.Backend)
     return
   }
   if (!(await isOnline())) {
-    logWarning('Version check failed, app is offline.')
+    logWarning('Version check failed, app is offline.', LogPrefix.Backend)
     return false
   }
   try {
@@ -86,7 +178,7 @@ async function checkForUpdates() {
     const currentVersion = app.getVersion()
     return semverGt(newVersion, currentVersion)
   } catch (error) {
-    logError('Could not check for new version of heroic')
+    logError('Could not check for new version of heroic', LogPrefix.Backend)
   }
 }
 
@@ -101,11 +193,11 @@ const showAboutWindow = () => {
   return app.showAboutPanel()
 }
 
-const handleExit = async () => {
-  const isLocked = existsSync(`${heroicGamesConfigPath}/lock`)
+async function handleExit(window: BrowserWindow) {
+  const isLocked = existsSync(join(heroicGamesConfigPath, 'lock'))
 
   if (isLocked) {
-    const { response } = await showMessageBox({
+    const { response } = await showMessageBox(window, {
       buttons: [i18next.t('box.no'), i18next.t('box.yes')],
       message: i18next.t(
         'box.quit.message',
@@ -117,7 +209,19 @@ const handleExit = async () => {
     if (response === 0) {
       return
     }
-    return app.exit()
+
+    // Kill all child processes
+    // This is very hacky
+    let killCommand = 'pkill --signal SIGINT '
+    if (isWindows) {
+      killCommand = 'Stop-Process -name '
+    }
+    const possibleChildren = ['legendary', 'gogdl']
+    possibleChildren.forEach(async (procName) => {
+      await execAsync(killCommand + procName).catch((error) => {
+        logInfo([`Unable to kill ${procName}, ignoring.`, error])
+      })
+    })
   }
   app.exit()
 }
@@ -169,18 +273,19 @@ type ErrorHandlerMessage = {
   logPath?: string
 }
 
-async function errorHandler({
-  error,
-  logPath
-}: ErrorHandlerMessage): Promise<void> {
+async function errorHandler(
+  { error, logPath }: ErrorHandlerMessage,
+  window?: BrowserWindow
+): Promise<void> {
   const noSpaceMsg = 'Not enough available disk space'
   const noCredentialsError = 'No saved credentials'
   if (logPath) {
     execAsync(`tail ${logPath} | grep 'disk space'`)
       .then(({ stdout }) => {
         if (stdout.includes(noSpaceMsg)) {
-          logError(noSpaceMsg)
-          return showErrorBox(
+          logError(noSpaceMsg, LogPrefix.Backend)
+          return showErrorBoxModal(
+            window,
             i18next.t('box.error.diskspace.title', 'No Space'),
             i18next.t(
               'box.error.diskspace.message',
@@ -189,11 +294,12 @@ async function errorHandler({
           )
         }
       })
-      .catch(() => logInfo('operation interrupted'))
+      .catch(() => logInfo('operation interrupted', LogPrefix.Backend))
   }
   if (error) {
     if (error.stderr.includes(noCredentialsError)) {
-      return showErrorBox(
+      return showErrorBoxModal(
+        window,
         i18next.t('box.error.credentials.title', 'Expired Credentials'),
         i18next.t(
           'box.error.credentials.message',
@@ -204,88 +310,16 @@ async function errorHandler({
   }
 }
 
-function genericErrorMessage(): void {
-  return showErrorBox(
-    i18next.t('box.error.generic.title', 'Unknown Error'),
-    i18next.t('box.error.generic.message', 'An Unknown Error has occurred')
-  )
-}
-
 function removeSpecialcharacters(text: string): string {
   const regexp = new RegExp('[:|/|*|?|<|>|\\|&|{|}|%|$|@|`|!|™|+]', 'gi')
   return text.replaceAll(regexp, '')
 }
 
 async function openUrlOrFile(url: string): Promise<string | void> {
-  if (process.platform === 'darwin') {
-    try {
-      await execAsync(`open ${url}`)
-    } catch (error) {
-      dialog.showErrorBox(
-        i18next.t('box.error.log.title', 'Log Not Found'),
-        i18next.t('box.error.log.message', 'No Log was found for this game')
-      )
-      return
-    }
-  }
-  if (process.platform === 'linux') {
-    try {
-      const fixedURL = url.replace('~', home)
-      await execAsync(`xdg-open '${fixedURL}'`)
-    } catch (error) {
-      dialog.showErrorBox(
-        i18next.t('box.error.log.title', 'Log Not Found'),
-        i18next.t('box.error.log.message', 'No Log was found for this game')
-      )
-      return
-    }
+  if (url.startsWith('http')) {
+    return shell.openExternal(url)
   }
   return shell.openPath(url)
-}
-
-/**
- * Checks given commands if they fullfil the given minimum version requirement.
- * @param commands      string list of commands to check.
- * @param version       minimum version to check against
- * @param all_fullfil   Can be set to false if only one command should fullfil
- *                      version requirement. (default: true)
- * @returns true if verrsion fullfil, else false
- */
-async function checkCommandVersion(
-  commands: string[],
-  version: string,
-  all_fullfil = true
-): Promise<boolean> {
-  let found = false
-  for (const command of commands) {
-    try {
-      const { stdout } = await execAsync(command + ' --version')
-      const commandVersion = stdout
-        ? stdout.match(/(\d+\.)(\d+\.)(\d+)/g)[0]
-        : null
-
-      if (semverGt(commandVersion, version) || commandVersion === version) {
-        logInfo(`Command '${command}' found. Version: '${commandVersion}'`)
-        if (!all_fullfil) {
-          return true
-        }
-        found = true
-      } else {
-        logWarning(
-          `Command ${command} version '${commandVersion}' not supported.`
-        )
-        if (all_fullfil) {
-          return false
-        }
-      }
-    } catch {
-      logWarning(`${command} command not found`)
-      if (all_fullfil) {
-        return false
-      }
-    }
-  }
-  return found
 }
 
 function clearCache() {
@@ -301,32 +335,107 @@ function clearCache() {
     cwd: 'lib-cache',
     name: 'gameinfo'
   })
+  const GOGapiInfoCache = new Store({
+    cwd: 'gog_store',
+    name: 'api_info_cache'
+  })
+  const GOGlibraryStore = new Store({ cwd: 'gog_store', name: 'library' })
+  GOGapiInfoCache.clear()
+  GOGlibraryStore.clear()
   installCache.clear()
   libraryCache.clear()
   gameInfoCache.clear()
 }
 
 function resetHeroic() {
-  const heroicFolder = `${app.getPath('appData')}/heroic`
-  rm(heroicFolder, { recursive: true, force: true }, () => {
+  const heroicFolders = [heroicGamesConfigPath, heroicConfigPath]
+  heroicFolders.forEach((folder) => {
+    rm(folder, { recursive: true, force: true }, () => null)
+  })
+  // wait a sec to avoid racing conditions
+  setTimeout(() => {
     app.relaunch()
     app.quit()
-  })
+  }, 1000)
+}
+
+function showItemInFolder(item: string) {
+  if (existsSync(item)) {
+    try {
+      shell.showItemInFolder(item)
+    } catch (error) {
+      logError(
+        `Failed to show item in folder with: ${error}`,
+        LogPrefix.Backend
+      )
+    }
+  }
+}
+
+function splitPathAndName(fullPath: string): { dir: string; bin: string } {
+  const dir = dirname(fullPath)
+  let bin = basename(fullPath)
+  // On Windows, you can just launch executables that are in the current working directory
+  // On Linux, you have to add a ./
+  if (!isWindows) {
+    bin = './' + bin
+  }
+  // Make sure to always return this as `dir, bin` to not break path
+  // resolution when using `join(...Object.values(...))`
+  return { dir, bin }
+}
+
+function getLegendaryBin(): { dir: string; bin: string } {
+  const settings = configStore.get('settings') as { altLeg: string }
+  if (settings?.altLeg) {
+    return splitPathAndName(settings.altLeg)
+  }
+  return splitPathAndName(
+    fixAsarPath(join(__dirname, 'bin', process.platform, 'legendary'))
+  )
+}
+
+function getGOGdlBin(): { dir: string; bin: string } {
+  const settings = configStore.get('settings') as { altGogdl: string }
+  if (settings?.altGogdl) {
+    return splitPathAndName(settings.altGogdl)
+  }
+  return splitPathAndName(
+    fixAsarPath(join(__dirname, 'bin', process.platform, 'gogdl'))
+  )
+}
+
+const specialCharactersRegex =
+  /('\w)|(\\(\w|\d){5})|(\\"(\\.|[^"])*")|[^((0-9)|(a-z)|(A-Z)|\s)]/g // addeed regex for capturings "'s" + unicodes + remove subtitles in quotes
+const cleanTitle = (title: string) =>
+  title
+    .replaceAll(specialCharactersRegex, '')
+    .replaceAll(' ', '-')
+    .replaceAll('®', '')
+    .toLowerCase()
+    .split('--definitive')[0]
+
+const formatEpicStoreUrl = (title: string) => {
+  const storeUrl = `https://www.epicgames.com/store/product/`
+  return `${storeUrl}${cleanTitle(title)}`
 }
 
 export {
-  checkCommandVersion,
   checkForUpdates,
   errorHandler,
   execAsync,
-  genericErrorMessage,
   handleExit,
   isOnline,
+  isEpicServiceOffline,
   openUrlOrFile,
   semverGt,
   showAboutWindow,
+  showItemInFolder,
   statAsync,
   removeSpecialcharacters,
   clearCache,
-  resetHeroic
+  resetHeroic,
+  getLegendaryBin,
+  getGOGdlBin,
+  formatEpicStoreUrl
 }
