@@ -1,55 +1,72 @@
-import React, { useContext, useEffect, useState } from 'react'
-import { useTranslation } from 'react-i18next'
-import { Link, NavLink, useLocation } from 'react-router-dom'
-import ElectronStore from 'electron-store'
-
-import cx from 'classnames'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import {
-  faStore,
-  faSlidersH,
   faBookOpen,
   faGamepad,
+  faSlidersH,
+  faStore,
   faUser
 } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import cx from 'classnames'
+import ElectronStore from 'electron-store'
+import React, { useCallback, useContext, useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import { NavLink, useHistory, useLocation } from 'react-router-dom'
+import { getAppSettings } from 'src/helpers'
+import ContextProvider from 'src/state/ContextProvider'
+import './index.css'
+
 const Store = window.require('electron-store')
 const configStore: ElectronStore = new Store({
   cwd: 'store'
 })
-
-import ContextProvider from 'src/state/ContextProvider'
-import './index.css'
-import { getAppSettings } from 'src/helpers'
+const gogStore: ElectronStore = new Store({
+  cwd: 'gog_store'
+})
 
 export default function SidebarLinks() {
   const { t } = useTranslation()
+  const history = useHistory()
   const [showUnrealMarket, setShowUnrealMarket] = useState(false)
 
   const { category, handleCategory, handleFilter } = useContext(ContextProvider)
 
   const location = useLocation() as { pathname: string }
   const isLibrary = location.pathname === '/'
-  const isLoggedIn = Boolean(configStore.get('userInfo'))
-  const showSubmenu = showUnrealMarket && isLibrary && isLoggedIn
+  const isStore = location.pathname.includes('store')
+  const isEpicLoggedIn = Boolean(configStore.get('userInfo'))
+  const isGOGLoggedIn = Boolean(gogStore.get('credentials'))
+  const showSidebar =
+    ((isEpicLoggedIn && isGOGLoggedIn) ||
+      (isEpicLoggedIn && showUnrealMarket)) &&
+    isLibrary
 
-  function toggleCategory(newCategory: string) {
-    if (category !== newCategory) {
-      handleCategory(newCategory)
-      handleFilter(newCategory === 'unreal' ? 'unreal' : 'all')
-    }
-  }
+  const pressAction = !isEpicLoggedIn && !isGOGLoggedIn ? '/login' : '/'
+  const displayIcon = isEpicLoggedIn || isGOGLoggedIn ? faGamepad : faUser
+
+  const toggleCategory = useCallback(
+    (newCategory: string) => {
+      if (category !== newCategory) {
+        handleCategory(newCategory)
+        handleFilter(newCategory === 'unreal' ? 'unreal' : 'all')
+      }
+    },
+    [category, handleCategory, handleFilter]
+  )
 
   useEffect(() => {
     getAppSettings().then(({ showUnrealMarket }) =>
       setShowUnrealMarket(showUnrealMarket)
     )
+    if (!isEpicLoggedIn && !isGOGLoggedIn) {
+      return history.push('/login')
+    }
   }, [])
 
   return (
-    <div className="Links">
+    <div className="SidebarLinks Sidebar__section">
       <NavLink
+        className="Sidebar__item"
         data-testid="library"
-        activeStyle={{ color: 'var(--secondary)', fontWeight: 500 }}
         isActive={(match, location) => {
           if (match) {
             return true
@@ -60,93 +77,110 @@ export default function SidebarLinks() {
           )
         }}
         exact
-        to="/"
+        to={pressAction}
       >
-        <FontAwesomeIcon
-          style={{ width: 'clamp(2vh, 25px, 30px)' }}
-          icon={isLoggedIn ? faGamepad : faUser}
-        />
-        {isLoggedIn ? t('Library') : t('button.login', 'Login')}
+        <div className="Sidebar__itemIcon">
+          <FontAwesomeIcon icon={displayIcon} />
+        </div>
+        {isEpicLoggedIn || isGOGLoggedIn
+          ? t('Library')
+          : t('button.login', 'Login')}
       </NavLink>
-      {showSubmenu && (
+      {showSidebar && (
         <>
-          <a
-            href="#"
-            onClick={() => toggleCategory('games')}
-            className={cx('subItem', { ['selected']: category === 'games' })}
-          >
-            {t('Epic Games', 'Epic Games')}
-          </a>
-          <a
-            href="#"
-            onClick={() => toggleCategory('unreal')}
-            className={cx('subItem', { ['selected']: category === 'unreal' })}
-          >
-            {t('Unreal Marketplace', 'Unreal Marketplace')}
-          </a>
-        </>
-      )}
-      {!isLoggedIn && (
-        <>
-          <Link
-            to="/"
-            className={cx('subItem', {
-              ['selected']: location.pathname === '/'
-            })}
-          >
-            {t('login.loginWithEpic', 'Login With Epic')}
-          </Link>
-          <Link
-            to="/login"
-            className={cx('subItem', {
-              ['selected']: location.pathname === '/login'
-            })}
-          >
-            {t('login.loginWithSid', 'Login with SID')}
-          </Link>
+          {isEpicLoggedIn && (
+            <a
+              href="#"
+              onClick={() => toggleCategory('epic')}
+              className={cx('Sidebar__item SidebarLinks__subItem', {
+                ['active']: category === 'epic'
+              })}
+            >
+              {t('Epic Games', 'Epic Games')}
+            </a>
+          )}
+          {isGOGLoggedIn && (
+            <a
+              href="#"
+              onClick={() => toggleCategory('gog')}
+              className={cx('Sidebar__item SidebarLinks__subItem', {
+                ['active']: category === 'gog'
+              })}
+            >
+              {t('GOG', 'GOG')}
+            </a>
+          )}
+          {showUnrealMarket && isEpicLoggedIn && (
+            <a
+              href="#"
+              onClick={() => toggleCategory('unreal')}
+              className={cx('Sidebar__item SidebarLinks__subItem', {
+                ['active']: category === 'unreal'
+              })}
+            >
+              {t('Unreal Marketplace', 'Unreal Marketplace')}
+            </a>
+          )}
         </>
       )}
       <NavLink
+        className="Sidebar__item"
         data-testid="settings"
-        activeStyle={{ color: 'var(--secondary)', fontWeight: 500 }}
         isActive={(match, location) => location.pathname.includes('settings')}
-        to={{
-          pathname: '/settings/default/general'
-        }}
+        to={{ pathname: '/settings/default/general' }}
       >
-        <FontAwesomeIcon
-          style={{ width: 'clamp(1vh, 22px, 28px)' }}
-          icon={faSlidersH}
-        />
-
+        <div className="Sidebar__itemIcon">
+          <FontAwesomeIcon icon={faSlidersH} />
+        </div>
         {t('Settings')}
       </NavLink>
       <NavLink
-        data-testid="store"
-        activeStyle={{ color: 'var(--secondary)', fontWeight: 500 }}
-        isActive={(match, location) => location.pathname.includes('epicstore')}
-        to={{
-          pathname: '/epicstore'
-        }}
+        className="Sidebar__item"
+        to={{ pathname: '/epicstore' }}
+        isActive={(match, location) => location.pathname.includes('store')}
       >
-        <FontAwesomeIcon
-          style={{ width: 'clamp(1vh, 22px, 28px)' }}
-          icon={faStore}
-        />
-        {t('store', 'Store')}
+        <div className="Sidebar__itemIcon">
+          <FontAwesomeIcon icon={faStore} />
+        </div>
+        {t('stores', 'Stores')}
       </NavLink>
+      {isStore && (
+        <>
+          <NavLink
+            data-testid="store"
+            className="Sidebar__item SidebarLinks__subItem"
+            isActive={(match, location) =>
+              location.pathname.includes('epicstore') ||
+              (location.pathname === '/store-page' &&
+                location.search.includes('epicgames.com/store'))
+            }
+            to={{ pathname: '/epicstore' }}
+          >
+            {t('store', 'Epic Store')}
+          </NavLink>
+          <NavLink
+            data-testid="store"
+            className="Sidebar__item SidebarLinks__subItem"
+            isActive={(match, location) =>
+              location.pathname.includes('gogstore') ||
+              (location.pathname === '/store-page' &&
+                location.search.includes('gog.com/en/game'))
+            }
+            to={{ pathname: '/gogstore' }}
+          >
+            {t('gog-store', 'GOG Store')}
+          </NavLink>
+        </>
+      )}
       <NavLink
         data-testid="wiki"
-        activeStyle={{ color: 'var(--secondary)', fontWeight: 500 }}
+        className="Sidebar__item"
         isActive={(match, location) => location.pathname.includes('wiki')}
-        to={{
-          pathname: '/wiki'
-        }}
+        to={{ pathname: '/wiki' }}
       >
-        <FontAwesomeIcon
-          style={{ width: 'clamp(1vh, 22px, 28px)' }}
-          icon={faBookOpen}
-        />
+        <div className="Sidebar__itemIcon">
+          <FontAwesomeIcon icon={faBookOpen} />
+        </div>
         {t('wiki', 'Wiki')}
       </NavLink>
     </div>
