@@ -1,14 +1,15 @@
 import { I18nextProvider, initReactI18next } from 'react-i18next'
 import HttpApi from 'i18next-http-backend'
-import LanguageDetector from 'i18next-browser-languagedetector'
 import React, { Suspense } from 'react'
 import ReactDOM from 'react-dom'
 import i18next from 'i18next'
+import { initGamepad } from './helpers/gamepad'
 
 import './index.css'
 import App from 'src/App'
 import GlobalState from 'src/state/GlobalState'
-import UpdateComponent from 'src/components/UI/UpdateComponent'
+import { UpdateComponentBase } from 'src/components/UI/UpdateComponent'
+import { initShortcuts } from './helpers/shortcuts'
 
 const Backend = new HttpApi(null, {
   addPath: 'build/locales/{{lng}}/{{ns}}',
@@ -16,13 +17,15 @@ const Backend = new HttpApi(null, {
   loadPath: 'locales/{{lng}}/{{ns}}.json'
 })
 
+initGamepad()
+initShortcuts()
+
 i18next
   // load translation using http -> see /public/locales
   // learn more: https://github.com/i18next/i18next-http-backend
   .use(Backend)
   // detect user language
   // learn more: https://github.com/i18next/i18next-browser-languageDetector
-  .use(LanguageDetector)
   .use(initReactI18next)
   .init({
     fallbackLng: 'en',
@@ -61,6 +64,8 @@ i18next
       'sv',
       'ta',
       'tr',
+      'uk',
+      'vi',
       'zh_Hans',
       'zh_Hant'
     ]
@@ -69,7 +74,7 @@ i18next
 ReactDOM.render(
   <React.StrictMode>
     <I18nextProvider i18n={i18next}>
-      <Suspense fallback={<UpdateComponent />}>
+      <Suspense fallback={<UpdateComponentBase message="Loading" />}>
         <GlobalState>
           <App />
         </GlobalState>
