@@ -1,11 +1,11 @@
 import './index.css'
 
 import React, {
-  Fragment,
   useContext,
   useEffect,
   useState,
-  MouseEvent
+  MouseEvent,
+  useMemo
 } from 'react'
 
 import { IpcRenderer } from 'electron'
@@ -29,13 +29,13 @@ import {
   GameInfo,
   GameStatus,
   InstallInfo,
-  InstallProgress
+  SavedInstallProgress
 } from 'src/types'
+import { useInstallProgress } from 'src/hooks'
 
 import GamePicture from '../GamePicture'
 import TimeContainer from '../TimeContainer'
 import prettyBytes from 'pretty-bytes'
-import { Checkbox } from '@material-ui/core'
 import GameRequirements from '../GameRequirements'
 import { GameSubMenu } from '..'
 import { InstallModal } from 'src/screens/Library/components'
@@ -60,8 +60,14 @@ export default function GamePage(): JSX.Element | null {
   const [tabToShow, setTabToShow] = useState('infoTab')
   const [showModal, setShowModal] = useState({ game: '', show: false })
 
-  const { libraryStatus, handleGameStatus, data, gameUpdates, platform } =
-    useContext(ContextProvider)
+  const {
+    libraryStatus,
+    handleGameStatus,
+    epicLibrary,
+    gogLibrary,
+    gameUpdates,
+    platform
+  } = useContext(ContextProvider)
   const gameStatus: GameStatus = libraryStatus.filter(
     (game: GameStatus) => game.appName === appName
   )[0]
@@ -69,22 +75,13 @@ export default function GamePage(): JSX.Element | null {
   const { status } = gameStatus || {}
   const previousProgress = JSON.parse(
     storage.getItem(appName) || '{}'
-  ) as InstallProgress
+  ) as SavedInstallProgress
 
   const [gameInfo, setGameInfo] = useState({} as GameInfo)
-  const [progress, setProgress] = useState(
-    previousProgress ??
-      ({
-        bytes: '0.00MiB',
-        eta: '00:00:00',
-        percent: '0.00%'
-      } as InstallProgress)
-  )
   const [autoSyncSaves, setAutoSyncSaves] = useState(false)
   const [savesPath, setSavesPath] = useState('')
   const [isSyncing, setIsSyncing] = useState(false)
   const [gameInstallInfo, setGameInstallInfo] = useState({} as InstallInfo)
-  const [installDlcs, setInstallDlcs] = useState(false)
   const [launchArguments, setLaunchArguments] = useState('')
   const [hasError, setHasError] = useState<{
     error: boolean
@@ -92,6 +89,8 @@ export default function GamePage(): JSX.Element | null {
   }>({ error: false, message: '' })
 
   const isWin = platform === 'win32'
+  const isMac = platform === 'darwin'
+  const isLinux = platform === 'linux'
   const isInstalling = status === 'installing'
   const isPlaying = status === 'playing'
   const isUpdating = status === 'updating'
@@ -106,10 +105,18 @@ export default function GamePage(): JSX.Element | null {
   useEffect(() => {
     const updateConfig = async () => {
       try {
-        const newInfo = await getGameInfo(appName)
+        let newInfo = await getGameInfo(appName, 'legendary')
+        if (!newInfo) {
+          newInfo = await getGameInfo(appName, 'gog')
+        }
         setGameInfo(newInfo)
-        getInstallInfo(appName)
-          .then((info) => setGameInstallInfo(info))
+        getInstallInfo(appName, newInfo.runner)
+          .then((info) => {
+            if (!info) {
+              throw 'Cannot get game info'
+            }
+            setGameInstallInfo(info)
+          })
           .catch((error) => {
             console.error(error)
             ipcRenderer.send('logError', error)
@@ -132,45 +139,19 @@ export default function GamePage(): JSX.Element | null {
       }
     }
     updateConfig()
-  }, [isInstalling, isPlaying, appName, data])
+  }, [isInstalling, isPlaying, appName, epicLibrary, gogLibrary])
 
-  useEffect(() => {
-    const progressInterval = setInterval(async () => {
-      if (isInstalling || isUpdating || isReparing) {
-        const progress: InstallProgress = await ipcRenderer.invoke(
-          'requestGameProgress',
-          appName
-        )
-
-        if (progress) {
-          if (previousProgress) {
-            const legendaryPercent = getProgress(progress)
-            const heroicPercent = getProgress(previousProgress)
-            const newPercent: number = Math.round(
-              (legendaryPercent / 100) * (100 - heroicPercent) + heroicPercent
-            )
-            progress.percent = `${newPercent}%`
-          }
-          return setProgress(progress)
-        }
-
-        return await handleGameStatus({
-          appName,
-          status
-        })
-      }
-    }, 1500)
-    return () => clearInterval(progressInterval)
-  }, [appName, isInstalling, isUpdating, isReparing])
+  const progressSince = useMemo(() => Date.now(), [gameStatus])
+  const progress = useInstallProgress(appName, gameInfo.runner, progressSince)
 
   async function handleUpdate() {
-    await handleGameStatus({ appName, status: 'updating' })
-    await updateGame(appName)
-    await handleGameStatus({ appName, status: 'done' })
-  }
-
-  function handleDlcs() {
-    setInstallDlcs(!installDlcs)
+    await handleGameStatus({
+      appName,
+      runner: gameInfo.runner,
+      status: 'updating'
+    })
+    await updateGame(appName, gameInfo.runner)
+    await handleGameStatus({ appName, runner: gameInfo.runner, status: 'done' })
   }
 
   function handleModal() {
@@ -181,6 +162,7 @@ export default function GamePage(): JSX.Element | null {
 
   if (gameInfo && gameInfo.install) {
     const {
+      runner,
       title,
       art_square,
       install: { install_path, install_size, version },
@@ -190,10 +172,10 @@ export default function GamePage(): JSX.Element | null {
       extra,
       developer,
       cloud_save_enabled,
-      canRunOffline
+      canRunOffline,
+      is_linux_native,
+      is_mac_native
     }: GameInfo = gameInfo
-    const haveDLCs = gameInstallInfo?.game?.owned_dlc?.length > 0
-    const DLCList = gameInstallInfo?.game?.owned_dlc
     const downloadSize =
       gameInstallInfo?.manifest?.download_size &&
       prettyBytes(Number(gameInstallInfo?.manifest?.download_size))
@@ -201,9 +183,13 @@ export default function GamePage(): JSX.Element | null {
       gameInstallInfo?.manifest?.disk_size &&
       prettyBytes(Number(gameInstallInfo?.manifest?.disk_size))
     const launchOptions = gameInstallInfo?.game?.launch_options || []
-    const pathname = isWin
-      ? `/settings/${appName}/other`
-      : `/settings/${appName}/wine`
+    // This should check for installed platform in the future
+    const isMacNative = isMac && is_mac_native
+    const isLinuxNative = isLinux && is_linux_native
+    const pathname =
+      isWin || isMacNative || isLinuxNative
+        ? `/settings/${appName}/other`
+        : `/settings/${appName}/wine`
 
     /*
     Other Keys:
@@ -231,12 +217,13 @@ export default function GamePage(): JSX.Element | null {
         {showModal.show && (
           <InstallModal
             appName={showModal.game}
+            runner={runner}
             backdropClick={() => setShowModal({ game: '', show: false })}
           />
         )}
         {title ? (
           <>
-            <GamePicture art_square={art_square} />
+            <GamePicture art_square={art_square} store={runner} />
             <div className={`gameTabs ${tabToShow}`}>
               {is_game && (
                 <>
@@ -290,29 +277,6 @@ export default function GamePage(): JSX.Element | null {
                             {t('game.installSize', 'Install Size')}:{' '}
                             {installSize ?? '...'}
                           </div>
-                          {haveDLCs && (
-                            <div className="itemContainer">
-                              <div className="dlcTitle">
-                                {t('dlc.title', 'DLCs')}
-                              </div>
-                              {DLCList.map(({ app_name, title }) => (
-                                <span key={app_name} className="dlcTitle">
-                                  {title}
-                                </span>
-                              ))}
-                              <span className="checkBox">
-                                <Checkbox
-                                  color="primary"
-                                  checked={installDlcs}
-                                  size="small"
-                                  onChange={() => handleDlcs()}
-                                />
-                                <span className="itemName">
-                                  {t('dlc.installDlcs', 'Install all DLCs')}
-                                </span>
-                              </span>
-                            </div>
-                          )}
                           <br />
                         </>
                       )}
@@ -395,7 +359,10 @@ export default function GamePage(): JSX.Element | null {
                       )}
                       {is_installed ? (
                         <Link
-                          to={{ pathname, state: { fromGameCard: false } }}
+                          to={{
+                            pathname,
+                            state: { fromGameCard: false, runner }
+                          }}
                           className={`button ${getButtonClass(is_installed)}`}
                         >
                           {`${getButtonLabel(is_installed)}`}
@@ -422,6 +389,8 @@ export default function GamePage(): JSX.Element | null {
                     appName={appName}
                     isInstalled={is_installed}
                     title={title}
+                    storeUrl={gameInfo.store_url}
+                    runner={gameInfo.runner}
                   />
                   <GameRequirements gameInfo={gameInfo} />
                 </>
@@ -457,8 +426,24 @@ export default function GamePage(): JSX.Element | null {
   function getInstallLabel(is_installed: boolean): React.ReactNode {
     const { eta, bytes, percent } = progress
 
+    let displayEta: string
+    const secondsLeft = eta
+    if (secondsLeft === Infinity) {
+      displayEta = ''
+    } else {
+      const seconds = `${secondsLeft % 60}`
+      const minutesLeft = Math.floor(secondsLeft / 60)
+      const minutes = `${minutesLeft % 60}`
+      const hours = `${Math.floor(minutesLeft / 60)}`
+      displayEta = [
+        hours.padStart(2, '0'),
+        minutes.padStart(2, '0'),
+        seconds.padStart(2, '0')
+      ].join(':')
+    }
+
     if (isReparing) {
-      return `${t('status.reparing')} ${percent ? `${percent}` : '...'}`
+      return `${t('status.reparing')} ${percent ? `${percent}%` : '...'}`
     }
 
     if (isMoving) {
@@ -466,13 +451,12 @@ export default function GamePage(): JSX.Element | null {
     }
 
     const currentProgress = `${
-      percent && bytes && eta ? `${percent} [${bytes}] | ETA: ${eta}` : '...'
+      percent && bytes
+        ? `${percent}% [${bytes}]${displayEta && ` | ETA: ${displayEta}`}`
+        : '...'
     }`
 
     if (isUpdating && is_installed) {
-      if (eta && eta.includes('verifying')) {
-        return `${t('status.reparing')}: ${percent} [${bytes}]`
-      }
       return `${t('status.updating')} ${currentProgress}`
     }
 
@@ -522,7 +506,7 @@ export default function GamePage(): JSX.Element | null {
   function handlePlay() {
     return async () => {
       if (status === 'playing' || status === 'updating') {
-        return sendKill(appName)
+        return sendKill(appName, gameInfo.runner)
       }
 
       if (autoSyncSaves) {
@@ -530,8 +514,7 @@ export default function GamePage(): JSX.Element | null {
         await syncSaves(savesPath, appName)
         setIsSyncing(false)
       }
-
-      await launch({ appName, t, launchArguments })
+      await launch({ appName, t, launchArguments, runner: gameInfo.runner })
 
       if (autoSyncSaves) {
         setIsSyncing(true)
@@ -562,7 +545,7 @@ export default function GamePage(): JSX.Element | null {
       previousProgress,
       progress,
       t,
-      installDlcs
+      runner: gameInfo.runner
     })
   }
 }
