@@ -1,6 +1,12 @@
 import React, { PureComponent } from 'react'
 
-import { GameInfo, GameStatus, RefreshOptions } from 'src/types'
+import {
+  GameInfo,
+  GameStatus,
+  InstalledInfo,
+  RefreshOptions,
+  WineVersionInfo
+} from 'src/types'
 import { TFunction, withTranslation } from 'react-i18next'
 import { getLegendaryConfig, getPlatform, install, launch } from 'src/helpers'
 import { i18n } from 'i18next'
@@ -19,7 +25,19 @@ const libraryStore: ElectronStore = new Store({
   cwd: 'lib-cache',
   name: 'library'
 })
+const wineDownloaderInfoStore: ElectronStore = new Store({
+  cwd: 'store',
+  name: 'wine-downloader-info'
+})
 
+const gogLibraryStore = new Store({ cwd: 'gog_store', name: 'library' })
+const gogInstalledGamesStore = new Store({
+  cwd: 'gog_store',
+  name: 'installed'
+})
+const gogConfigStore = new Store({
+  cwd: 'gog_store'
+})
 const RTL_LANGUAGES = ['fa']
 
 type T = TFunction<'gamepage'> & TFunction<'translations'>
@@ -32,7 +50,9 @@ interface Props {
 
 interface StateProps {
   category: string
-  data: GameInfo[]
+  epicLibrary: GameInfo[]
+  gogLibrary: GameInfo[]
+  wineVersions: WineVersionInfo[]
   error: boolean
   filter: string
   filterText: string
@@ -46,10 +66,31 @@ interface StateProps {
 }
 
 export class GlobalState extends PureComponent<Props> {
+  loadGOGLibrary = (): Array<GameInfo> => {
+    const games = gogLibraryStore.has('games')
+      ? (gogLibraryStore.get('games') as GameInfo[])
+      : []
+    const installedGames =
+      (gogInstalledGamesStore.get('installed') as Array<InstalledInfo>) || []
+    for (const igame in games) {
+      for (const installedGame in installedGames) {
+        if (installedGames[installedGame].appName == games[igame].app_name) {
+          games[igame].install = installedGames[installedGame]
+          games[igame].is_installed = true
+        }
+      }
+    }
+
+    return games
+  }
   state: StateProps = {
-    category: 'games',
-    data: libraryStore.has('library')
+    category: 'epic',
+    epicLibrary: libraryStore.has('library')
       ? (libraryStore.get('library') as GameInfo[])
+      : [],
+    gogLibrary: this.loadGOGLibrary(),
+    wineVersions: wineDownloaderInfoStore.has('wine-releases')
+      ? (wineDownloaderInfoStore.get('wine-releases') as WineVersionInfo[])
       : [],
     error: false,
     filter: 'all',
@@ -66,17 +107,18 @@ export class GlobalState extends PureComponent<Props> {
   refresh = async (checkUpdates?: boolean): Promise<void> => {
     let updates = this.state.gameUpdates
     console.log('refreshing')
-    const currentLibraryLength = this.state.data?.length
-    let library: Array<GameInfo> =
+    const currentLibraryLength = this.state.epicLibrary?.length
+    let epicLibrary: Array<GameInfo> =
       (libraryStore.get('library') as Array<GameInfo>) || []
 
-    if (!library.length || !this.state.data.length) {
+    const gogLibrary: Array<GameInfo> = this.loadGOGLibrary()
+    if (!epicLibrary.length || !this.state.epicLibrary.length) {
       ipcRenderer.send(
         'logInfo',
         'No cache found, getting data from legendary...'
       )
       const { library: legendaryLibrary } = await getLegendaryConfig()
-      library = legendaryLibrary
+      epicLibrary = legendaryLibrary
     }
 
     try {
@@ -89,12 +131,13 @@ export class GlobalState extends PureComponent<Props> {
 
     this.setState({
       filterText: '',
-      data: library,
+      epicLibrary,
+      gogLibrary,
       gameUpdates: updates,
       refreshing: false
     })
 
-    if (currentLibraryLength !== library.length) {
+    if (currentLibraryLength !== epicLibrary.length) {
       ipcRenderer.send('logInfo', 'Force Update')
       this.forceUpdate()
     }
@@ -113,6 +156,26 @@ export class GlobalState extends PureComponent<Props> {
       ipcRenderer.send('logError', error)
     }
     this.refresh(checkForUpdates)
+  }
+
+  refreshWineVersionInfo = async (fetch: boolean): Promise<void> => {
+    if (this.state.platform !== 'linux') {
+      return
+    }
+    ipcRenderer.send('logInfo', 'Refreshing wine downloader releases')
+    try {
+      this.setState({ refreshing: true })
+      const releases = await ipcRenderer.invoke('refreshWineVersionInfo', fetch)
+      this.setState({
+        wineVersions: releases,
+        refreshing: false
+      })
+    } catch (error) {
+      this.setState({ refreshing: false })
+      console.error(error)
+      ipcRenderer.send('logError', error)
+      ipcRenderer.send('logError', 'Refreshing wine downloader releases failed')
+    }
   }
 
   handleSearch = (input: string) => this.setState({ filterText: input })
@@ -179,9 +242,15 @@ export class GlobalState extends PureComponent<Props> {
   filterPlatform = (library: GameInfo[], filter: string) => {
     switch (filter) {
       case 'win':
-        return library.filter((game) => !game.is_mac_native)
+        return library.filter((game) =>
+          process.platform == 'darwin'
+            ? !game.is_mac_native
+            : !game.is_linux_native
+        )
       case 'mac':
         return library.filter((game) => game.is_mac_native)
+      case 'linux':
+        return library.filter((game) => game.is_linux_native)
       default:
         return library.filter((game) => game.is_game)
     }
@@ -191,7 +260,8 @@ export class GlobalState extends PureComponent<Props> {
     appName,
     status,
     folder,
-    progress
+    progress,
+    runner
   }: GameStatus) => {
     const { libraryStatus, gameUpdates } = this.state
     const currentApp =
@@ -230,7 +300,11 @@ export class GlobalState extends PureComponent<Props> {
         libraryStatus: updatedLibraryStatus
       })
 
-      return this.refreshLibrary({})
+      // This waits for backend to synchronize installed games (GOG)
+      setTimeout(() => {
+        this.refreshLibrary({})
+      }, 500)
+      return
     }
 
     if (currentApp && currentApp.status === 'updating' && status === 'done') {
@@ -271,7 +345,7 @@ export class GlobalState extends PureComponent<Props> {
         (game) => game.appName !== appName
       )
       this.setState({ libraryStatus: updatedLibraryStatus })
-      ipcRenderer.send('removeShortcut', appName)
+      ipcRenderer.send('removeShortcut', appName, runner)
 
       return this.refreshLibrary({})
     }
@@ -321,15 +395,16 @@ export class GlobalState extends PureComponent<Props> {
 
   async componentDidMount() {
     const { i18n, t } = this.props
-    const { data, gameUpdates = [], libraryStatus } = this.state
+    const { epicLibrary, gameUpdates = [], libraryStatus } = this.state
 
     // Deals launching from protocol. Also checks if the game is already running
-    ipcRenderer.on('launchGame', async (e, appName) => {
+    ipcRenderer.on('launchGame', async (e, appName, runner) => {
       const currentApp = libraryStatus.filter(
         (game) => game.appName === appName
       )[0]
       if (!currentApp) {
-        return launch({ appName, t })
+        // Add finding a runner for games
+        return launch({ appName, t, runner })
       }
     })
 
@@ -337,7 +412,7 @@ export class GlobalState extends PureComponent<Props> {
       const currentApp = libraryStatus.filter(
         (game) => game.appName === appName
       )[0]
-      const { appName, installPath } = args
+      const { appName, installPath, runner } = args
       if (!currentApp || (currentApp && currentApp.status !== 'installing')) {
         return install({
           appName,
@@ -346,11 +421,10 @@ export class GlobalState extends PureComponent<Props> {
           isInstalling: false,
           previousProgress: null,
           progress: {
-            bytes: '0.00MiB',
-            eta: '00:00:00',
-            percent: '0.00%'
+            percent: 0
           },
-          t
+          t,
+          runner
         })
       }
     })
@@ -359,15 +433,15 @@ export class GlobalState extends PureComponent<Props> {
       const { libraryStatus } = this.state
       this.handleGameStatus({ ...libraryStatus, ...args })
     })
-
-    const user = configStore.get('userInfo')
+    const legendaryUser = Boolean(configStore.get('userInfo'))
+    const gogUser = Boolean(gogConfigStore.get('userData'))
     const platform = await getPlatform()
-    const category = storage.getItem('category') || 'games'
+    const category = storage.getItem('category') || 'epic'
     const filter = storage.getItem('filter') || 'all'
     const layout = storage.getItem('layout') || 'grid'
     const language = storage.getItem('language') || 'en'
 
-    if (!user) {
+    if (!legendaryUser) {
       await ipcRenderer.invoke('getUserInfo')
     }
 
@@ -379,11 +453,11 @@ export class GlobalState extends PureComponent<Props> {
     i18n.changeLanguage(language)
     this.setState({ category, filter, language, layout, platform })
 
-    if (user) {
+    if (legendaryUser || gogUser) {
       this.refreshLibrary({
         checkForUpdates: true,
         fullRefresh: true,
-        runInBackground: Boolean(data.length)
+        runInBackground: Boolean(epicLibrary.length)
       })
     }
 
@@ -412,8 +486,17 @@ export class GlobalState extends PureComponent<Props> {
 
   render() {
     const { children } = this.props
-    const { data, filterText, filter, platform, filterPlatform } = this.state
-    let filteredLibrary = data
+    const {
+      epicLibrary,
+      wineVersions,
+      gogLibrary,
+      filterText,
+      filter,
+      platform,
+      filterPlatform
+    } = this.state
+    let filteredEpicLibrary = epicLibrary
+    let filteredGOGLibrary = gogLibrary
     const language = storage.getItem('language') || 'en'
     const isRTL = RTL_LANGUAGES.includes(language)
 
@@ -421,19 +504,34 @@ export class GlobalState extends PureComponent<Props> {
       const filterRegex = new RegExp(filterText, 'i')
       const textFilter = ({ title, app_name }: GameInfo) =>
         filterRegex.test(title) || filterRegex.test(app_name)
-      filteredLibrary = this.filterPlatform(
-        this.filterLibrary(data, filter).filter(textFilter),
+      filteredEpicLibrary = this.filterPlatform(
+        this.filterLibrary(epicLibrary, filter).filter(textFilter),
+        filterPlatform
+      )
+      filteredGOGLibrary = this.filterPlatform(
+        this.filterLibrary(gogLibrary, filter).filter(textFilter),
         filterPlatform
       )
     } catch (error) {
       console.log(error)
     }
 
+    let recentGames: GameInfo[] = []
+
+    if (epicLibrary.length > 0) {
+      recentGames = [...getRecentGames(epicLibrary)]
+    }
+    if (gogLibrary.length > 0) {
+      recentGames = [...recentGames, ...getRecentGames(gogLibrary)]
+    }
+
     return (
       <ContextProvider.Provider
         value={{
           ...this.state,
-          data: filteredLibrary,
+          epicLibrary: filteredEpicLibrary,
+          gogLibrary: filteredGOGLibrary,
+          wineVersions: wineVersions,
           handleCategory: this.handleCategory,
           handleFilter: this.handleFilter,
           handleGameStatus: this.handleGameStatus,
@@ -444,7 +542,8 @@ export class GlobalState extends PureComponent<Props> {
           platform: platform,
           refresh: this.refresh,
           refreshLibrary: this.refreshLibrary,
-          recentGames: getRecentGames(data)
+          refreshWineVersionInfo: this.refreshWineVersionInfo,
+          recentGames
         }}
       >
         {children}
